+++ packages/dictionary/src/ReadonlyDictionary.ts
@@ -1,10 +1,10 @@
 import { BinaryPredicate, Cloneable, Mapping } from '@jamashita/anden-type';
-import { Collection } from '@jamashita/lluvia-collection';
+import { Collection, NarrowingBinaryPredicate } from '@jamashita/lluvia-collection';
 
 export interface ReadonlyDictionary<out K, out V> extends Collection<K, V>, Cloneable<ReadonlyDictionary<K, V>> {
   filter(predicate: BinaryPredicate<V, K>): ReadonlyDictionary<K, V>;
 
-  filter<W extends V>(predicate: BinaryPredicate<W, K>): ReadonlyDictionary<K, W>;
+  filter<W extends V>(predicate: NarrowingBinaryPredicate<V, W, K>): ReadonlyDictionary<K, W>;
 
   has(key: K): boolean;