+++ packages/sequence/src/Sequence.ts
@@ -1,4 +1,5 @@
 import { BinaryPredicate, Mapping } from '@jamashita/anden-type';
+import { NarrowingBinaryPredicate } from '@jamashita/lluvia-collection';
 import { ReadonlySequence } from './ReadonlySequence';
 
 export interface Sequence<out V> extends ReadonlySequence<V> {
@@ -6,7 +7,7 @@ export interface Sequence<out V> extends ReadonlySequence<V> {
 
   filter(predicate: BinaryPredicate<V, number>): Sequence<V>;
 
-  filter<W extends V>(predicate: BinaryPredicate<W, number>): Sequence<W>;
+  filter<W extends V>(predicate: NarrowingBinaryPredicate<V, W, number>): Sequence<W>;
 
   map<W>(mapping: Mapping<V, W>): Sequence<W>;